Move getSvgProps function into it's own file
