+++ lib/octicons_react/script/build.js
@@ -41,57 +41,9 @@ function writeIcons(file) {
   const count = icons.length
   const code = `${GENERATED_HEADER}
 import React from 'react'
+import getSvgProps from '../get-svg-props'
 
 ${icons.map(({code}) => code).join('\n')}
-
-function closestNaturalHeight(naturalHeights, height) {
-  return naturalHeights
-    .map(naturalHeight => parseInt(naturalHeight, 10))
-    .reduce(
-      (acc, naturalHeight) => (naturalHeight <= height ? naturalHeight : acc),
-      naturalHeights[0]
-    )
-}
-
-const sizeMap = {
-  small: 16,
-  medium: 32,
-  large: 64
-}
-
-function getSvgProps({
-  ariaLabel,
-  className,
-  size,
-  verticalAlign,
-  svgDataByHeight
-}) {
-  const height = sizeMap[size] || size
-  const naturalHeight = closestNaturalHeight(
-    Object.keys(svgDataByHeight),
-    height
-  )
-  const naturalWidth = svgDataByHeight[naturalHeight].width
-  const width = height * (naturalWidth / naturalHeight)
-  const path = svgDataByHeight[naturalHeight].path
-
-  return {
-    'aria-hidden': ariaLabel ? 'false' : 'true',
-    'aria-label': ariaLabel,
-    role: 'img',
-    className,
-    viewBox: \`0 0 $\{naturalWidth} $\{naturalHeight}\`,
-    width,
-    height,
-    fill: 'currentColor',
-    style: {
-      display: 'inline-block',
-      userSelect: 'none',
-      verticalAlign
-    },
-    dangerouslySetInnerHTML: { __html: path }
-  }
-}
 
 const iconsByName = {
   ${icons.map(({key, name}) => `'${key}': ${name}`).join(',\n  ')}
